% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{calc_mean_queue}
\alias{calc_mean_queue}
\title{Calculate the time-weighted mean queue length.}
\usage{
calc_mean_queue(arrivals, simulation_end_time, groups = NULL)
}
\arguments{
\item{arrivals}{Dataframe with times for each patient with each resource.}

\item{simulation_end_time}{Time at end of simulation run.}

\item{groups}{Optional list of columns to group by for the calculation.}
}
\value{
Tibble with column containing mean queue length.
}
\description{
Calculate the time-weighted mean queue length.
}
