% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{calc_mean_time_in_system}
\alias{calc_mean_time_in_system}
\title{Calculate the mean time in the system for finished patients.}
\usage{
calc_mean_time_in_system(arrivals, groups = NULL)
}
\arguments{
\item{arrivals}{Dataframe with times for each patient with each resource.}

\item{groups}{Optional list of columns to group by for the calculation.}
}
\value{
Tibble with column containing the mean time in the system.
}
\description{
Calculate the mean time in the system for finished patients.
}
