% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{calc_utilisation}
\alias{calc_utilisation}
\title{Calculate the resource utilisation}
\usage{
calc_utilisation(
  resources,
  simulation_end_time,
  groups = NULL,
  summarise = TRUE
)
}
\arguments{
\item{resources}{Dataframe with times patients use or queue for resources.}

\item{simulation_end_time}{Time at end of simulation run.}

\item{groups}{Optional list of columns to group by for the calculation.}

\item{summarise}{If TRUE, return overall utilisation. If FALSE, just return
the resource dataframe with the additional columns interval_duration,
effective_capacity and utilisation.}
}
\value{
Tibble with columns containing result for each resource.
}
\description{
Utilisation is given by the total effective usage time (`in_use`) over
the total time intervals considered (`dt`).
}
\details{
Credit: The utilisation calculation is adapted from the
`plot.resources.utilization()` function in simmer.plot 0.1.18, which is
shared under an MIT Licence (Ucar I, Smeets B (2023). simmer.plot: Plotting
Methods for 'simmer'. https://r-simmer.org
https://github.com/r-simmer/simmer.plot.).
}
