% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{confidence_interval_method}
\alias{confidence_interval_method}
\title{Use the confidence interval method to select the number of replications.}
\usage{
confidence_interval_method(
  replications,
  desired_precision,
  metric,
  yaxis_title,
  path,
  min_rep = NULL
)
}
\arguments{
\item{replications}{Number of times to run the model.}

\item{desired_precision}{Desired mean deviation from confidence interval.}

\item{metric}{Name of performance metric to assess.}

\item{yaxis_title}{Label for y axis.}

\item{path}{Path inc. filename to save figure to.}

\item{min_rep}{A suggested minimum number of replications (default=NULL).}
}
\value{
Dataframe with results from each replication.
}
\description{
Use the confidence interval method to select the number of replications.
}
