% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{create_replication_tabuliser}
\alias{create_replication_tabuliser}
\title{Create a Replication Tabuliser}
\usage{
create_replication_tabuliser()
}
\value{
A list of functions with the following methods:
  - `tab_update(state)`: Record results from Welford tracker state
  - `summary_table()`: Get accumulated results as a dataframe
  - `get_data_points()`: Get raw data points
  - `get_cumulative_means()`: Get running means
  - `get_std_devs()`: Get standard deviations
  - `get_lcis()`: Get lower confidence bounds
  - `get_ucis()`: Get upper confidence bounds
  - `get_deviations()`: Get precision deviations
}
\description{
Observes and records results from Welford statistics tracker.
Updates each time new data is processed. Can generate a results dataframe.

This function is based on the Python class `ReplicationTabulizer` from Tom
Monks (2021) sim-tools: fundamental tools to support the simulation process
in python (https://github.com/sim-tools/sim-tools) (MIT Licence).
}
