% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\title{Generate a list of default parameters}
\usage{
defaults()
}
\value{
A list containing the default parameters:

- `patient_inter`: Average interval (in minutes) between patient arrivals.
- `mean_n_consult_time`: Mean consultation time per patient (in minutes).
- `number_of_nurses`: Number of available nurses.
- `verbose`: Whether to print logs of when patients arrive and use resources.
- `data_collection_period`: Total period (in minutes) for data collection.
- `number_of_runs`: Number of simulation runs.
- `cores`: Number of cores for computation (1 = sequential, -1 = all
  available, otherwise specify e.g. 2, 3, 4...).
}
\description{
This function returns a list containing the default parameters for the model.
}
