% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_warmup.R
\name{get_interval_means}
\alias{get_interval_means}
\title{Find the mean results per interval in each replication in the results.}
\usage{
get_interval_means(result, interval_size)
}
\arguments{
\item{result}{Named list with `arrivals` and `resources` containing results
from the run of a simulation for the desired number of replications.}

\item{interval_size}{Size of batch intervals.}
}
\value{
List containing results from each interval and replication for
each metric.
}
\description{
Find the mean results per interval in each replication in the results.
}
