% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{get_run_results}
\alias{get_run_results}
\title{Get average results for the provided single run.}
\usage{
get_run_results(results, run_number, simulation_end_time)
}
\arguments{
\item{results}{Named list with `arrivals` containing output from
`get_mon_arrivals()` and `resources` containing output from
`get_mon_resources()` (`per_resource = TRUE` and `ongoing = TRUE`).}

\item{run_number}{Integer representing index of current simulation run.}

\item{simulation_end_time}{Time at end of simulation run.}
}
\value{
Tibble with processed results from replication.
}
\description{
Get average results for the provided single run.
}
