% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{klimit}
\alias{klimit}
\title{Calculate the klimit}
\usage{
klimit(look_ahead, n)
}
\arguments{
\item{look_ahead}{Minimum additional replications to look ahead to assess
stability of precision.}

\item{n}{Number of replications performed so far.}
}
\value{
Number of additional replications to verify stability (integer).
}
\description{
Determines the number of additional replications to check after precision is
reached, scaling with total replications if they are greater than 100.
Rounded down to nearest integer.
}
