% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run simulation}
\usage{
model(run_number, param_class)
}
\arguments{
\item{run_number}{Integer representing index of current simulation run.}

\item{param_class}{Instance of Defaults containing model parameters.}
}
\value{
Named list with two tables: monitored arrivals and resources
}
\description{
Run simulation
}
