% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Create a named list of default model parameters (which can be altered).}
\usage{
parameters(
  patient_inter = 4L,
  mean_n_consult_time = 10L,
  number_of_nurses = 5L,
  warm_up_period = 1440L * 7L,
  data_collection_period = 1440L * 14L,
  number_of_runs = 25L,
  scenario_name = NULL,
  cores = 1L,
  seed_offset = 0L,
  log_to_console = FALSE,
  log_to_file = FALSE,
  file_path = NULL
)
}
\arguments{
\item{patient_inter}{Mean inter-arrival time between patients in minutes.}

\item{mean_n_consult_time}{Mean nurse consultation time in minutes.}

\item{number_of_nurses}{Number of available nurses (int).}

\item{warm_up_period}{Duration of warm-up period in minutes.}

\item{data_collection_period}{Duration of data collection period in
minutes.}

\item{number_of_runs}{Number of simulation runs (int).}

\item{scenario_name}{Label for the scenario (int|float|string).}

\item{cores}{Number of cores to use for parallel execution (int).}

\item{seed_offset}{Optional value to add to each run number when generating
seeds for simulation replications (used for sensitivity analyses).}

\item{log_to_console}{Whether to print activity log to console.}

\item{log_to_file}{Whether to save activity log to file.}

\item{file_path}{Path to save log to file.}
}
\value{
A named list containing the parameters for the model.
}
\description{
When input to model(), valid_inputs() will fetch the inputs to this
function and compare them against the provided list, to ensure no
new keys have been add to the list.
}
