% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_replications.R
\name{process_replications}
\alias{process_replications}
\title{Process results from each replication.}
\usage{
process_replications(env)
}
\arguments{
\item{env}{Simmer Environment or list of simmer Environments.}
}
\value{
Tibble with results from each replication.
}
\description{
For each replication (there can be one or many), calculate the number of
arrivals, as well as mean results for each resource: wait time, activity
time and utilisation. Credit: The utilisation calculation is taken from
the plot.resources.utilization() function in simmer.plot 0.1.18, which is
shared under an MIT Licence (Ucar I, Smeets B (2023). simmer.plot: Plotting
Methods for 'simmer'. https://r-simmer.org
https://github.com/r-simmer/simmer.plot.).
}
