% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{replications_algorithm}
\alias{replications_algorithm}
\title{Replication algorithm to automatically select number of replications.}
\usage{
replications_algorithm(
  metrics = c("mean_waiting_time_nurse", "mean_serve_time_nurse", "utilisation_nurse"),
  param = parameters(),
  desired_precision = 0.05,
  initial_replications = 3L,
  look_ahead = 5L,
  replication_budget = 1000L
)
}
\arguments{
\item{metrics}{List of performance measure to track (should correspond to
column names from the run results dataframe).}

\item{param}{Model parameters (from parameters()).}

\item{desired_precision}{The target half width precision for the algorithm
(i.e. percentage deviation of the confidence interval from the mean,
expressed as a proportion, e.g. 0.05 = 5\%).}

\item{initial_replications}{Number of initial replications to perform.
Note that the minimum solution will be the value of initial_replications
(i.e. if require 20 initial replications but was resolved in 5, solution
output will still be 20). Although, if initial_replications < 3, solution
will still be at least 3, as that is the minimum required to calculate the
confidence intervals.}

\item{look_ahead}{Minimum additional replications to look ahead to assess
stability of precision. When the number of replications is <= 100, the
value of look_ahead is used. When they are > 100, then
look_ahead / 100 * max(n, 100) is used.}

\item{replication_budget}{Maximum allowed replications. Use for larger
models where replication runtime is a constraint.}
}
\value{
Named list with `nreps` (a named list with minimum number of
replications required to meet the precision for each metric) and
`summary_table` (dataframe containing cumulative statistics for each
replication for each metric).
}
\description{
Implements an adaptive replication algorithm for selecting the
appropriate number of simulation replications based on statistical
precision.

Uses the "Replications Algorithm" from Hoad, Robinson, & Davies (2010).
Automated selection of the number of replications for a discrete-event
simulation. Journal of the Operational Research Society.
https://www.jstor.org/stable/40926090.

Given a model's performance measure and a user-set confidence interval
half width prevision, automatically select the number of replications.
Combines the "confidence intervals" method with a sequential look-ahead
procedure to determine if a desired precision in the confidence interval
is maintained.

This class is based on the Python class `ReplicationsAlgorithm` from Tom
Monks (2021) sim-tools: fundamental tools to support the simulation process
in python (https://github.com/TomMonks/sim-tools) (MIT Licence).
}
