% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\docType{class}
\name{ReplicationTabuliser}
\alias{ReplicationTabuliser}
\title{Observes and records results from WelfordStats.}
\value{
Object of `R6Class` with methods for storing and tabulising results.
}
\description{
Updates each time new data is processed. Can generate a results dataframe.

This class is based on the Python class `ReplicationTabulizer` from Tom
Monks (2021) sim-tools: fundamental tools to support the simulation process
in python (https://github.com/sim-tools/sim-tools) (MIT Licence).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_points}}{List containing each data point.}

\item{\code{cumulative_mean}}{List of the running mean.}

\item{\code{std}}{List of the standard deviation.}

\item{\code{lci}}{List of the lower confidence interval bound.}

\item{\code{uci}}{List of the upper confidence interval bound.}

\item{\code{deviation}}{List of the percentage deviation of the confidence
interval half width from the mean.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReplicationTabuliser-update}{\code{ReplicationTabuliser$update()}}
\item \href{#method-ReplicationTabuliser-summary_table}{\code{ReplicationTabuliser$summary_table()}}
\item \href{#method-ReplicationTabuliser-clone}{\code{ReplicationTabuliser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationTabuliser-update"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationTabuliser-update}{}}}
\subsection{Method \code{update()}}{
Add new results from WelfordStats to the appropriate lists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationTabuliser$update(stats)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stats}}{An instance of WelfordStats containing updated statistical
measures like the mean, standard deviation and confidence intervals.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationTabuliser-summary_table"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationTabuliser-summary_table}{}}}
\subsection{Method \code{summary_table()}}{
Creates a results table from the stored lists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationTabuliser$summary_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Stored results compiled into a dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationTabuliser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationTabuliser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationTabuliser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
