% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{runner}
\alias{runner}
\title{Run simulation for multiple replications, sequentially or in parallel.}
\usage{
runner(param)
}
\arguments{
\item{param}{Named list of model parameters.}
}
\value{
Named list with three tables: monitored arrivals, monitored
resources, and the processed results from each run.
}
\description{
Run simulation for multiple replications, sequentially or in parallel.
}
