% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{trial}
\alias{trial}
\title{Run simulation for multiple replications, sequentially or in parallel.}
\usage{
trial(param)
}
\arguments{
\item{param}{List containing parameters for the simulations.}
}
\value{
Named list with two tables: monitored arrivals and resources.
}
\description{
Note: The parallel processing is implemented using `parLapply` because
`mcLapply`` does not work on Windows and `future_lapply`` would often get
stuck.
}
