% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{valid_inputs}
\alias{valid_inputs}
\title{Check validity of input parameters}
\usage{
valid_inputs(run_number, param)
}
\arguments{
\item{run_number}{Integer representing index of current simulation run.}

\item{param}{List containing parameters for the simulation.}
}
\description{
Check validity of input parameters
}
