% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\docType{class}
\name{WelfordStats}
\alias{WelfordStats}
\title{Computes running sample mean and variance using Welford's algorithm.}
\value{
Object of `R6Class` with methods for running mean and variance
calculation.
}
\description{
They are computed via updates to a stored value, rather than storing lots of
data and repeatedly taking the mean after new values have been added.

Implements Welford's algorithm for updating mean and variance.
See Knuth. D `The Art of Computer Programming` Vol 2. 2nd ed. Page 216.

This class is based on the Python class `OnlineStatistics` from Tom Monks
(2021) sim-tools: fundamental tools to support the simulation process in
python (https://github.com/sim-tools/sim-tools) (MIT Licence).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n}}{Number of observations.}

\item{\code{latest_data}}{Latest data point.}

\item{\code{mean}}{Running mean.}

\item{\code{sq}}{Running sum of squares of differences.}

\item{\code{alpha}}{Significance level for confidence interval calculations.
For example, if alpha is 0.05, then the confidence level is 95\%.}

\item{\code{observer}}{Observer to notify on updates.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WelfordStats-new}{\code{WelfordStats$new()}}
\item \href{#method-WelfordStats-update}{\code{WelfordStats$update()}}
\item \href{#method-WelfordStats-variance}{\code{WelfordStats$variance()}}
\item \href{#method-WelfordStats-std}{\code{WelfordStats$std()}}
\item \href{#method-WelfordStats-std_error}{\code{WelfordStats$std_error()}}
\item \href{#method-WelfordStats-half_width}{\code{WelfordStats$half_width()}}
\item \href{#method-WelfordStats-lci}{\code{WelfordStats$lci()}}
\item \href{#method-WelfordStats-uci}{\code{WelfordStats$uci()}}
\item \href{#method-WelfordStats-deviation}{\code{WelfordStats$deviation()}}
\item \href{#method-WelfordStats-clone}{\code{WelfordStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-new"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-new}{}}}
\subsection{Method \code{new()}}{
Initialise the WelfordStats object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$new(data = NULL, alpha = 0.05, observer = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Initial data sample.}

\item{\code{alpha}}{Significance level for confidence interval calculations.}

\item{\code{observer}}{Observer to notify on updates.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `WelfordStats` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-update"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-update}{}}}
\subsection{Method \code{update()}}{
Update running statistics with a new data point.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$update(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A new data point.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-variance"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-variance}{}}}
\subsection{Method \code{variance()}}{
Computes the variance of the data points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-std"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-std}{}}}
\subsection{Method \code{std()}}{
Computes the standard deviation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$std()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-std_error"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-std_error}{}}}
\subsection{Method \code{std_error()}}{
Computes the standard error of the mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$std_error()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-half_width"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-half_width}{}}}
\subsection{Method \code{half_width()}}{
Computes the half-width of the confidence interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$half_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-lci"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-lci}{}}}
\subsection{Method \code{lci()}}{
Computes the lower confidence interval bound.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$lci()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-uci"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-uci}{}}}
\subsection{Method \code{uci()}}{
Computes the upper confidence interval bound.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$uci()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-deviation"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-deviation}{}}}
\subsection{Method \code{deviation()}}{
Computes the precision of the confidence interval expressed
as the percentage deviation of the half width from the mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$deviation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WelfordStats-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WelfordStats-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WelfordStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
