% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{create_replications_algorithm}
\alias{create_replications_algorithm}
\title{Replication Algorithm for Automatic Selection}
\usage{
create_replications_algorithm(
  param,
  metrics,
  desired_precision = 0.1,
  initial_replications = 3L,
  look_ahead = 5L,
  replication_budget = 1000L,
  verbose = TRUE
)
}
\arguments{
\item{param}{Model parameters (list).}

\item{metrics}{List of performance measure names to track (should correspond
to column names from the run results dataframe).}

\item{desired_precision}{Target half width precision for the algorithm
(i.e. percentage deviation of the confidence interval from the mean,
expressed as a proportion, e.g. 0.1 = 10%). Choice is fairly arbitrary.}

\item{initial_replications}{Number of initial replications to perform
(default: 3).}

\item{look_ahead}{Minimum additional replications to look ahead to assess
stability of precision. When replications are <= 100, the value of
look_ahead is used. When > 100, then look_ahead / 100 * max(n, 100)
is used (default: 5).}

\item{replication_budget}{Maximum allowed replications. Use for larger
models where replication runtime is a constraint (default: 1000).}

\item{verbose}{Boolean, whether to print messages about parameters
(default: TRUE).}
}
\value{
A list of functions with the following methods:
  - `select()`: Execute the replication algorithm
  - `get_nreps()`: Get minimum replications required for each metric
  - `get_summary_table()`: Get full results table with cumulative
    statistics for each replication
  - `get_reps()`: Get number of replications performed
}
\description{
Implements an adaptive replication algorithm for selecting the
appropriate number of simulation replications based on statistical
precision.

Uses the "Replications Algorithm" from Hoad, Robinson, & Davies (2010).
Automated selection of the number of replications for a discrete-event
simulation. Journal of the Operational Research Society.
https://www.jstor.org/stable/40926090.

Given a model's performance measure and a user-set confidence interval
half width precision, automatically select the number of replications.
Combines the "confidence intervals" method with a sequential look-ahead
procedure to determine if a desired precision in the confidence interval
is maintained.

This function is based on the Python class `ReplicationsAlgorithm` from Tom
Monks (2021) sim-tools: fundamental tools to support the simulation process
in python (https://github.com/sim-tools/sim-tools) (MIT Licence).
}
