% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{get_run_results}
\alias{get_run_results}
\title{Process the raw monitored arrivals and resources.}
\usage{
get_run_results(results, run_number)
}
\arguments{
\item{results}{Named list with `arrivals` containing output from
`get_mon_arrivals()` and `resources` containing output from
`get_mon_resources()` (`per_resource = TRUE` and `ongoing = TRUE`).}

\item{run_number}{Integer representing index of current simulation run.}
}
\value{
Tibble with processed results from replication.
}
\description{
For the provided replication, calculate the:
(1) number of arrivals
(2) mean wait time for each resource
(3) mean activity time for each resource
(4) mean resource utilisation.
}
\details{
Credit: The utilisation calculation is taken from the
`plot.resources.utilization()` function in simmer.plot 0.1.18, which is
shared under an MIT Licence (Ucar I, Smeets B (2023). simmer.plot: Plotting
Methods for 'simmer'. https://r-simmer.org
https://github.com/r-simmer/simmer.plot.).

Note: When calculating the mean wait time, it is rounded to 10 decimal
places. This is to resolve an issue that occurs because `start_time`,
`end_time` and `activity_time` are all to 14 decimal places, but the
calculations can produce tiny negative values due to floating-point errors.
}
