% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{valid_inputs}
\title{Run simulation}
\usage{
model(run_number, param)

valid_inputs(run_number, param)
}
\arguments{
\item{run_number}{Integer representing index of current simulation run.}

\item{param}{List containing parameters for the simulation.}
}
\value{
Named list with two tables: monitored arrivals and resources
}
\description{
Run simulation

Check validity of inputs to `model()`
}
