% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{runner}
\alias{runner}
\title{Run simulation for multiple replications, sequentially or in parallel.}
\usage{
runner(param)
}
\arguments{
\item{param}{Named list of model parameters.}
}
\value{
Named list with two tables: monitored arrivals and resources.
}
\description{
Run simulation for multiple replications, sequentially or in parallel.
}
