% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{trial}
\alias{trial}
\title{Run simulation for multiple replications, sequentially or in parallel.}
\usage{
trial(param_class)
}
\arguments{
\item{param_class}{Instance of Defaults containing model parameters.}
}
\value{
Named list with two tables: monitored arrivals and resources.
}
\description{
Run simulation for multiple replications, sequentially or in parallel.
}
